feat: some minor refactor
